Update navigation & pages to properly display with icon directives
- Switched to <icon file=""></icon> for all icons
- Updated image references

In local environment on Windows, icon directive showed image path rather
than image. This patch reconstructs the menu to use the icon directive
used on upstream, but with a line on the icon-provider.js that allows the
directive to work on local environments as well.

This allows for consistency with upstream requirements and more seamless
updates when preparing UI changes for upstream

Tested: Viewed pages with icons to ensure they loaded properly

Change-Id: If654ecf2862e91698fc572f1dffa71efcff16b75
Signed-off-by: Kathy Pine <kathryn.elainex.pine@intel.com>