Fix loader issue in hardware status page
When tried to get the memory and pcie device info via
redfish, Its giving empty array. This condition is not
checked in webui which is causing the loader to spin always.
Checked the members.count value before process the request
and load the page with data.

Tested:
Navigated to hardware status page and its loaded properly
when there is empty response from redfish.

Change-Id: Ie0988d75a24291e26b3e6388fc640d0a69c18187
Signed-off-by: Mohammad Asadullah <mohammad.asadullahx@intel.com>