OpenBmc Power operation reboot issue
When click to start power operation then toaster message
show with wrong information, then changed method where
each n every specific interval hitting API to get power
operation information to check whether it is On or Off
and then show in the toaster message with response, same
process done for reboot and shutdown process.

Tested:
Checked with power on and off to get response from
API and show message of correct information in the
power operation, reboot and shutdown process.

Change-Id: If04b738fb42d13746ba67204ebc1f5a7b6f3eb43
Signed-off-by: Mohammad Asadullah <mohammad.asadullahx@intel.com>