Event log API request parameter wrong with -value
Event log API request sending with - value in the parameter
like skip=-25&top=25, so fixed negative parameter and
put top value 25 for first time and skip=0 and hit API in reverse,
now get last data as first data from the response of API and
then first pagination will be added in the list of logs, and
again hit API with top=200 and get second last data as second data
from the response and added in the second pagination and so on.

Tested:
Tested with 2500 logs in the network, now its sending request
with correct value like, skip=2475&top=25, skip=2275&top=200 and so on.

Change-Id: Ic32ff6da79c1a5891b596ddd8190a683fea0678a
Signed-off-by: Mohammad Asadullah <mohammad.asadullahx@intel.com>