mTLS authentication improvement for webui.
This change allows webui to properly handle mTLS authentication.
The login box is only shown for a moment before redirection to the
main page. As the first requests in the webui (root directory,js
and css files) are whitelisted in bmcweb, the authorization flow
with cookies is not used. This caused webui to display a login box
when the connection was authenticated. Making an additional REST
API call allows using the route providing the UI right away.

Tested:
Manually verified redirection.
Automatic test with selenium succeeded.

Change-Id: Ib614de8df49759e145a0933144053726f3bb759f
Signed-off-by: Golgowski, Wiktor <wiktor.golgowski@intel.com>